docs: move swagger folder from docs to /swagger
