+++ swagger/template/index.tpl
